requirements.txt
# Twitter Bot Requirements

# Twitter API
tweepy>=4.14.0

# Google Gemini API
google-genai>=1.0.0

# OpenAI API (optional)
openai>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Optional: For Jupyter notebooks (development only)
jupyter>=1.0.0
ipykernel>=6.0.0
